Move type from schema to types
