+++ packages/signed-api/src/schema.ts
@@ -70,17 +70,6 @@ export const configSchema = z.strictObject({
 
 export type Config = z.infer<typeof configSchema>;
 
-// TODO: Move to types
-export type InternalSignedData = {
-  airnode: string;
-  templateId: string;
-  beaconId: string;
-  timestamp: string;
-  encodedValue: string;
-  signature: string;
-  isOevBeacon: boolean;
-};
-
 export const envBooleanSchema = z.union([z.literal('true'), z.literal('false')]).transform((val) => val === 'true');
 
 // We apply default values to make it convenient to omit certain environment variables. The default values should be
